Style is no longer exposed via Presentation
* Require Gaphas 2.1.1 (master) code
* Everything styling related to gaphor.core.styling.
* Provide style data through update/draw contexts
* Style is requested via the diagram object
* Use Style type all around
