+++ gaphor/UML/actions/tests/test_flow.py
@@ -1,7 +1,8 @@
 from gaphas.canvas import instant_cairo_context
 
 import gaphor.UML as UML
-from gaphor.diagram.shapes import DrawContext
+from gaphor.core.modeling import DrawContext
+from gaphor.core.styling import DEFAULT_STYLE
 from gaphor.tests.testcase import TestCase
 from gaphor.UML.actions.flow import FlowItem
 
@@ -41,11 +42,11 @@ class FlowTestCase(TestCase):
         flow = self.create(FlowItem, UML.ControlFlow)
         context = DrawContext(
             cairo=instant_cairo_context(),
+            style=DEFAULT_STYLE,
             hovered=True,
             focused=True,
             selected=True,
             dropzone=False,
-            style={},
         )
 
         flow.draw(context)
