+++ gaphor/UML/classes/tests/test_class.py
@@ -2,10 +2,9 @@
 Test classes.
 """
 
-from gaphas.canvas import instant_cairo_context
-
 from gaphor import UML
-from gaphor.diagram.shapes import SizeContext
+from gaphor.core.modeling import UpdateContext
+from gaphor.core.styling import DEFAULT_STYLE
 from gaphor.tests.testcase import TestCase
 from gaphor.UML.classes.klass import ClassItem
 
@@ -15,7 +14,7 @@ def compartments(item):
 
 
 def context():
-    return SizeContext(cairo=instant_cairo_context(), style={})
+    return UpdateContext(style=DEFAULT_STYLE)
 
 
 class ClassTestCase(TestCase):
