+++ gaphor/UML/interactions/tests/test_executionspecification.py
@@ -1,7 +1,8 @@
 from gaphas.canvas import instant_cairo_context
 
 from gaphor import UML
-from gaphor.diagram.shapes import DrawContext
+from gaphor.core.modeling import DrawContext
+from gaphor.core.styling import DEFAULT_STYLE
 from gaphor.diagram.tests.fixtures import allow, connect, disconnect
 from gaphor.UML.interactions.executionspecification import ExecutionSpecificationItem
 from gaphor.UML.interactions.lifeline import LifelineItem
@@ -25,11 +26,11 @@ def test_draw_on_canvas(diagram):
     exec_spec.draw(
         DrawContext(
             cairo=cr,
+            style=DEFAULT_STYLE,
             selected=False,
             focused=False,
             hovered=False,
             dropzone=False,
-            style={},
         )
     )
