+++ gaphor/core/modeling/__init__.py
@@ -4,7 +4,8 @@ from gaphor.core.modeling.coremodel import (
     NamedElement,
     PackageableElement,
 )
-from gaphor.core.modeling.diagram import Diagram
+from gaphor.core.modeling.diagram import Diagram, DrawContext, UpdateContext
 from gaphor.core.modeling.elementfactory import ElementFactory
 from gaphor.core.modeling.event import *
-from gaphor.core.modeling.presentation import Presentation, StyleSheet
+from gaphor.core.modeling.presentation import Presentation
+from gaphor.core.modeling.stylesheet import StyleSheet
