+++ gaphor/core/styling/tests/test_properties.py
