+++ gaphor/diagram/tests/test_shapes.py
@@ -2,6 +2,7 @@ import cairo
 import pytest
 from gaphas.geometry import Rectangle
 
+from gaphor.core.styling import DEFAULT_STYLE
 from gaphor.diagram.shapes import (
     Box,
     DrawContext,
@@ -29,11 +30,11 @@ def context():
     cr = cairo.Context(surface)
     return DrawContext(
         cairo=cr,
+        style=DEFAULT_STYLE,
         selected=False,
         focused=False,
         hovered=False,
         dropzone=False,
-        style={},
     )
