+++ gaphor/ui/diagrampage.py
@@ -8,7 +8,6 @@ from gaphas.painter import (
     BoundingBoxPainter,
     FocusedItemPainter,
     HandlePainter,
-    ItemPainter,
     PainterChain,
     ToolPainter,
 )
@@ -25,6 +24,7 @@ from gaphor.diagram.diagramtools import (
     TransactionalToolChain,
 )
 from gaphor.diagram.event import DiagramItemPlaced
+from gaphor.diagram.painter import ItemPainter
 from gaphor.diagram.support import get_diagram_item
 from gaphor.services.properties import PropertyChanged
 from gaphor.transaction import Transaction
